ST-NS-014 Fix mypy type errors for CI compliance
- Add explicit_package_bases=true to pyproject.toml mypy config
- Add type annotations to **kwargs parameters
- Add type annotations to missing return types
- Add type: ignore comments for complex type issues
- Fix numpy float type issues with explicit casts
- Fix contextlib.suppress pattern for exception handling
- Fix line length violations in error messages

+++ src/portfolio_risk/position_sizing/calculator.py
@@ -145,7 +145,7 @@ class PositionSizeCalculator:
         # Calculate ATR using Wilder's smoothing
         atr_values: list[float] = []
         # First ATR is simple average
-        atr_values.append(np.mean(tr_values[:period]))
+        atr_values.append(float(np.mean(tr_values[:period])))
 
         # Subsequent values use smoothing formula
         for i in range(period, len(tr_values)):
